requirements.txt
fastapi==0.70.0
numpy==1.23.1
# onnxruntime==1.9.0
opencv-python==4.6.0.66
# openshift==0.12.1
Pillow==9.2.0
# pydicom==2.2.2
uvicorn==0.15.0
# wheel==0.37.0
pandas==1.5.3
python-multipart==0.0.5
# pylibjpeg==1.3.0
# pylibjpeg-libjpeg==1.2.0
# pylibjpeg-openjpeg==1.1.1
# pylibjpeg-rle==1.1.0
# pynetdicom==2.0.1
# xgboost==0.90 #1.6.0
# scikit-learn==1.0.2

# opencv-python-headless==4.5.5.62
# streamlit
# watchdog==2.1.6
aiofiles==0.8.0
# python-gdcm==3.0.10
# scipy
# imageio==2.14.1
matplotlib==3.5.2

scikit-learn==1.1.1
tensorflow==2.9.1
# keras==2.9.0
efficientnet==1.1.1
torch==1.12.0
torchvision==0.13.0
ttach==0.0.3
captum==0.7.0
grad-cam==1.4.6